add zsh

+++ setup.sh
@@ -4,7 +4,11 @@
 dotfiles=(.*)
 for f in ${dotfiles[*]}; do
     if [ ${#f} -gt 2 ] && [ "$f" !=  ".git" ]; then
-        rm ~/$f
+        # Remove the file if it exists. 
+        if [ -e ~/$f ]; then
+            rm ~/$f
+        fi
+        # Create a symbolic link to the file in the home directory.
         ln -sv $(pwd)/$f ~
     fi
 done
